Create project setup with make, cmake, conan, and docker
The root Makefile wraps the tasks to be executed, like compilation and
test execution. As a build system generator, absent uses CMake following
a modern practice based on targets and properties. To manage dependencies,
absent relies on conan which now adds GoogleTest to exercise the library
utilities. Finally, in order to have reproducible and isolated
environments, especially aiming CI pipelines, absent provides a suitable
Dockerfile to compile and run the tests using gcc 8.