change: Don't implement combinators in terms of a common syntax
Trying to combine the API for optional-like and either made the code
much more complex than it needs to be, for both API and implementation.

Moreover, the differences between the templates had required
hacks that didn't scale well given the lack of a sane way to do
partial applications in the type-level.

Now, the main combinators and each adapter shall have its own API and
implementation, not sharing a common syntax, which wasn't common at all.

Furthermore, some API renaming took place to make usage simpler and more obvious.

Unfortunately, this is a breaking-change and clients must need to be adapted.