chore: Use noexcept operator to declare noexcept combinators
Even so, client code might just map types into types and usually
not fail with an exception, but rather with a nullable. It is safer
to assume that client code may throw when called from
inside a combinator. Thus this patch uses the noexcept operator to
reduce the likelihood of raising issues when the client code does throw.